Notification/function restructure
