+++ functions/logOut.js
@@ -1,0 +1,11 @@
+const logOut = (fb, props) => {
+        
+          
+        fb.signOut().then(() => {
+            props.navigation.navigate({routeName: 'Index'});
+          }).catch((error) => {
+            console.log(error.toString())
+          }); 
+}
+
+export default logOut;
