+++ screens/MapScreen.js
@@ -6,6 +6,7 @@ import * as Location from 'expo-location';
 import * as firebase from 'firebase';
 import { MaterialCommunityIcons } from '@expo/vector-icons'; 
 import { FontAwesome } from '@expo/vector-icons'; 
+import addVid from '../functions/addvid';
 
 
 
@@ -19,61 +20,6 @@ const MapScreen = props => {
   const db = firebase.firestore();
   var user = firebase.auth().currentUser;
 
-
-
-
-  //make external component
-  addVid = () => {
-
-    try{
-        db.collection('Assistance Request').where('request_ID', '==', addRequest).get()
-        .then(snapshot => {
-          snapshot.forEach(doc => {
-            const data = doc.data();
-            const docID = doc.id;
-            console.log(docID);
-            console.log(data);
-
-            db.collection("Assistance Request").doc(doc.id).update({
-                vid: user.uid,
-                status: 'Assigned to a Volunteer'
-      
-              }),
-
-                db.collection("users").doc(data.uid).get().then(snapshot => {
-                    const userInfo = snapshot.data();
-                    console.log(userInfo);
-
-                    fetch('https://exp.host/--/api/v2/push/send', {
-                    method: 'POST',
-                    headers: {
-                        Accept: 'application/json',
-                        'Accept-Encoding': 'gzip, deflate',
-                        'Content-Type': 'application/json',
-                    },
-                    body: JSON.stringify({
-                        to: userInfo.expoToken,
-                        data: { extraData: 'Request ID: ' + data.request_ID},
-                        title: 'Request made on ' + data.date + ' was picked up' ,
-                        body: 'Please check your assistance request of type:'  + '\r\n' + data.request_Type ,
-                    }),
-                    });
-            
-                    
-              })
-
-          });
-       
-       
-        })
-
-    }
-    catch(error){
-        console.log(error.toString())
-      }
-
-
-}
 
   useEffect(() => {
     db.collection('Assistance Request')
@@ -167,7 +113,7 @@ const MapScreen = props => {
             onChangeText={(addRequest) => setAddRequest(addRequest)}
             />
             </View>
-            <TouchableOpacity style={styles.Btn} onPress = {addVid}>
+            <TouchableOpacity style={styles.Btn} onPress = {() => addVid(addRequest, db, user)}>
              <Text style={styles.subText} >Add request</Text>
 
         </TouchableOpacity>
