+++ screens/SignupScreen.js
@@ -1,6 +1,7 @@
 import React, {useState} from 'react';
 import { StyleSheet, Text, View, TextInput, Button, TouchableOpacity, Image} from 'react-native';
 import * as firebase from 'firebase';
+import signUpUser from '../functions/signUpUser'
 
 const SignupScreen = props => {
 
@@ -10,70 +11,8 @@ const SignupScreen = props => {
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
   const [passwordRepeat, setPasswordRepeat] = useState('');
-  
+  const fb = firebase.auth();
 
-  signUpUser = () => {
-    try{
-      if(name == '' ){
-      alert(' Name is empty!')
-            return;
-    }
-    else{
-      if(number == ''){
-        alert('Number Field is empty!')
-        return;
-    }
-    else{
-      if(email == ''){
-        alert('Email Field is empty!')
-        return;
-    }
-    else{
-      if(password == ''){
-        alert('Password Field is empty!')
-        return;
-    }
-    else{
-      if(password.length<6){
-        alert('Please enter a password more than 6 characters!')
-            return;
-    }
-    else{
-      if(passwordRepeat == password){
-
-        firebase.auth().createUserWithEmailAndPassword(email, password).then(function (user) {
-          console.log(user),
-          props.navigation.navigate({routeName: 'Home'});
-          db.collection("users").doc(user.user.uid).set({
-            name: name,
-            number: number,
-            email: email
-          })
-  
-      });
-
-      }
-      else{
-
-        alert('Your passwords do not match!')
-        
-      }
-     
-    }
-      
-    }
-      
-    }
-
-    }
-
-    }
-        
-    }  
-    catch(error){
-        console.log(error.toString())
-    }
-}
 
   return(
     <View style={styles.screen} >
@@ -125,7 +64,7 @@ const SignupScreen = props => {
     </View>
 
     <View>
-    <TouchableOpacity style={styles.signBtn}  onPress = {signUpUser}>
+    <TouchableOpacity style={styles.signBtn}  onPress = {() => signUpUser(name, number, email, password, passwordRepeat, fb, props, db)}>
       <Text style={styles.signText}>Sign Up</Text>
     </TouchableOpacity>
     </View>
